refactor image upload for products

+++ modules/product/product.service.js
@@ -231,6 +231,9 @@ class ProductsService {
     primary,
     { _id: adminId }
   ) {
+    const matchPrimaryInUpload = filesToUpload.filter(
+      item => item.large === productData.images[0].primary.large
+    );
     productData.images = {
       primary: {
         large: LARGE_SAD_BACKPACK,
@@ -264,12 +267,13 @@ class ProductsService {
       if (primary?.large) {
         productData.images.primary = primary;
       } else {
-        await uploadService.deleteFiles(
-          Object.values(product.images.primary).filter(
-            item => typeof item === 'string'
-          )
-        );
-        const uploadResult = await uploadService.uploadFiles(primary);
+        if (!matchPrimaryInUpload.length)
+          await uploadService.deleteFiles(
+            Object.values(product.images.primary).filter(
+              item => typeof item === 'string'
+            )
+          );
+        const uploadResult = await uploadService.uploadFiles([primary]);
         const imagesResults = await uploadResult[0];
         productData.images.primary = imagesResults?.fileNames;
       }
